+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Full changelog
 ==============
 
+0.4 (unreleased)
+----------------
+
+* Added the ability to add arbitrary callbacks to ``CallbackDict`` and
+ ``CallbackList`` via the ``.callbacks`` attribute. [#25]
+
 0.3 (2020-05-04)
 ----------------